requirements.txt
# PDF处理
PyMuPDF>=1.23.0  # fitz, PDF解析
pdf2image>=1.16.0  # PDF转图片(需要poppler)

# 图像处理
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0  # CPU版本,如需GPU用paddlepaddle-gpu

# 深度学习(可选,Stage 2+)
# torch>=2.0.0
# torchvision>=0.15.0
# ultralytics>=8.0.0  # YOLOv8

# 数据处理
pandas>=2.0.0
scikit-learn>=1.3.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 开发工具
jupyter>=1.0.0
ipykernel>=6.25.0
tqdm>=4.65.0

# 实用工具
python-dotenv>=1.0.0
loguru>=0.7.0
